fix: web bugs (#49)
* fix: seed bug

* fix: line share bug

* fix: remote login toast

* chore: code refactoring

* chore: update react-table version

* fix: route only with projectId

* fix: reset page in issue table
