+++ apps/web/src/containers/tables/FeedbackTable/EditableCell/index.ts
@@ -13,4 +13,4 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-export { default } from './FeedbackSearch';
+export { default } from './EditableCell';
