+++ apps/web/src/containers/tables/FeedbackTable/FeedbackDeleteDialog/index.ts
@@ -13,20 +13,4 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-interface IProps extends React.PropsWithChildren {
-  colSpan: number;
-}
-
-const TableLoadingRow: React.FC<IProps> = ({ colSpan }) => {
-  return (
-    <tr>
-      <td colSpan={colSpan} style={{ padding: 0, height: 0 }}>
-        <div className="relative w-full bg-gray-200 rounded">
-          <div className="top-0 h-1 rounded w-full relative loading" />
-        </div>
-      </td>
-    </tr>
-  );
-};
-
-export default TableLoadingRow;
+export { default } from './FeedbackDeleteDialog';
