+++ apps/web/src/pages/_app.css
@@ -34,7 +34,7 @@
 
 body {
   @apply font-12-regular text-primary bg-primary;
-  font-family: "LINESeedEN", sans-serif;
+  font-family: LINESeedJP, LINESeedKR, LINESeedEN, sans-serif;
 }
 
 a {
@@ -61,28 +61,79 @@ th {
   font-family: 'LINESeedEN';
   font-style: thin;
   font-weight: 100;
-  src: url('/fonts/LINESeedEN/EOT/LINESeedEN-Th.otf') format('opentype'),
-    url('/fonts/LINESeedEN/EOT/LINESeedEN-Th.eot') format('embedded-opentype'),
-    url('/fonts/LINESeedEN/WOFF2/LINESeedEN-Th.woff2') format('woff2'),
-    url('/fonts/LINESeedEN/WOFF/LINESeedEN-Th.woff') format('woff');
+  src: local(※), url('/fonts/LINESeedEN/WOFF2/LINESeedEN-Th.woff2') format('woff2'),
+    url('/fonts/LINESeedEN/WOFF2/LINESeedEN-Th.woff') format('woff'),
+    url('/fonts/LINESeedEN/WOFF/LINESeedEN-Th.ttf') format('truetype');
 }
 
 @font-face {
   font-family: 'LINESeedEN';
   font-style: normal;
   font-weight: 400;
-  src: url('/fonts/LINESeedEN/EOT/LINESeedEN-Rg.otf') format('opentype'),
-    url('/fonts/LINESeedEN/EOT/LINESeedEN-Rg.eot') format('embedded-opentype'),
-    url('/fonts/LINESeedEN/WOFF2/LINESeedEN-Rg.woff2') format('woff2'),
-    url('/fonts/LINESeedEN/WOFF/LINESeedEN-Rg.woff') format('woff');
+  src: local(※), url('/fonts/LINESeedEN/WOFF2/LINESeedEN-Rg.woff2') format('woff2'),
+    url('/fonts/LINESeedEN/WOFF2/LINESeedEN-Rg.woff') format('woff'),
+    url('/fonts/LINESeedEN/WOFF/LINESeedEN-Rg.ttf') format('truetype');
 }
 
 @font-face {
   font-family: 'LINESeedEN';
   font-style: bold;
   font-weight: 700;
-  src: url('/fonts/LINESeedEN/EOT/LINESeedEN-Bd.otf') format('opentype'),
-    url('/fonts/LINESeedEN/EOT/LINESeedEN-Bd.eot') format('embedded-opentype'),
-    url('/fonts/LINESeedEN/WOFF2/LINESeedEN-Bd.woff2') format('woff2'),
-    url('/fonts/LINESeedEN/WOFF/LINESeedEN-Bd.woff') format('woff');
+  src: local(※), url('/fonts/LINESeedEN/WOFF2/LINESeedEN-Bd.woff2') format('woff2'),
+    url('/fonts/LINESeedEN/WOFF2/LINESeedEN-Bd.woff') format('woff'),
+    url('/fonts/LINESeedEN/WOFF/LINESeedEN-Bd.ttf') format('truetype');
+}
+
+@font-face {
+  font-family: 'LINESeedJP';
+  font-style: thin;
+  font-weight: 100;
+  src: local(※), url('/fonts/LINESeedJP/WOFF2/LINESeedJP-Th.woff2') format('woff2'),
+    url('/fonts/LINESeedJP/WOFF2/LINESeedJP-Th.woff') format('woff'),
+    url('/fonts/LINESeedJP/WOFF/LINESeedJP-Th.ttf') format('truetype');
+}
+
+@font-face {
+  font-family: 'LINESeedJP';
+  font-style: normal;
+  font-weight: 400;
+  src: local(※), url('/fonts/LINESeedJP/WOFF2/LINESeedJP-Rg.woff2') format('woff2'),
+    url('/fonts/LINESeedJP/WOFF2/LINESeedJP-Rg.woff') format('woff'),
+    url('/fonts/LINESeedJP/WOFF/LINESeedJP-Rg.ttf') format('truetype');
+}
+
+@font-face {
+  font-family: 'LINESeedJP';
+  font-style: bold;
+  font-weight: 700;
+  src: local(※), url('/fonts/LINESeedJP/WOFF2/LINESeedJP-Bd.woff2') format('woff2'),
+    url('/fonts/LINESeedJP/WOFF2/LINESeedJP-Bd.woff') format('woff'),
+    url('/fonts/LINESeedJP/WOFF/LINESeedJP-Bd.ttf') format('truetype');
+}
+
+@font-face {
+  font-family: 'LINESeedKR';
+  font-style: thin;
+  font-weight: 100;
+  src: local(※), url('/fonts/LINESeedKR/WOFF2/LINESeedKR-Th.woff2') format('woff2'),
+    url('/fonts/LINESeedKR/WOFF2/LINESeedKR-Th.woff') format('woff'),
+    url('/fonts/LINESeedKR/WOFF/LINESeedKR-Th.ttf') format('truetype');
+}
+
+@font-face {
+  font-family: 'LINESeedKR';
+  font-style: normal;
+  font-weight: 400;
+  src: local(※), url('/fonts/LINESeedKR/WOFF2/LINESeedKR-Rg.woff2') format('woff2'),
+    url('/fonts/LINESeedKR/WOFF2/LINESeedKR-Rg.woff') format('woff'),
+    url('/fonts/LINESeedKR/WOFF/LINESeedKR-Rg.ttf') format('truetype');
+}
+
+@font-face {
+  font-family: 'LINESeedKR';
+  font-style: bold;
+  font-weight: 700;
+  src: local(※), url('/fonts/LINESeedKR/WOFF2/LINESeedKR-Bd.woff2') format('woff2'),
+    url('/fonts/LINESeedKR/WOFF2/LINESeedKR-Bd.woff') format('woff'),
+    url('/fonts/LINESeedKR/WOFF/LINESeedKR-Bd.ttf') format('truetype');
 }
